Moves model logic to a separate file * The logic that pertains to model atmospheres is now in phoebe/atmospheres/models.py. * Model atmosphere instantiation from fits files is moved to a class method from_path(). * Class attributes of ModelAtmosphere and its subclasses now suffice for passband handling. * A master list of model atmospheres is in _atmtable. * pb.save() has been updated to use the new ModelAtmosphere classes.
