+++ phoebe/atmospheres/models.py
@@ -1,0 +1,366 @@
+import numpy as np
+import glob
+import os
+import re
+
+
+class ModelAtmosphere:
+    """
+    A parent class for handling model atmosphere data. Please note that only
+    derived classes should be instantiated.
+
+    Model atmospheres are approximations to stellar atmospheres. Each model
+    connects input parameters (for example, effective temperature, surface
+    gravity, and chemical abundances) to output parameters (for example,
+    intensities at different wavelengths and angles). The ModelAtmosphere
+    class provides a common interface for handling different model
+    atmospheres.
+
+    In order to have a model atmosphere supported, the following attributes
+    need to be defined:
+
+    * `name` (string): name of the model atmosphere
+    * `prefix` (string): prefix for the model atmosphere fits keywords
+    * `basic_axis_names` (list): names of the basic axes; basic axes are
+        axes that span the basic n-dimensional model atmosphere grid. The grid
+        can be sparsely populated, but it must be regular.
+    * `associated_axis_names` (list): names of the associated axes; associated
+        axes are axes that are not part of the basic grid, but are associated
+        with it. For example, the mus axis is associated with the basic grid
+        of teffs, loggs, and abuns. Associated axes must be fully defined in
+        all basic grid nodes.
+    * `mus` (array): specific angles, mu=cos(theta), where theta is the angle
+        between the observer and the surface normal.
+    * `wls` (array): wavelengths of the model atmosphere intensities.
+    * `units` (float): intensity unit conversion factor. The intensities are
+        usually given in erg/s/cm^2/A, which is converted to W/m^3 by
+        multiplying with this factor.
+
+    In addition, the following methods need to or may be overloaded:
+
+    * `parse_rules`: provides rules for parsing atmosphere fits filenames to
+        extract basic axis values.
+    * `limb_treatment`: defines how intensities at the exact limb (mu=0) should
+        be treated. By default, the intensities are linearly extrapolated to
+        mu=0.
+
+    When a model atmosphere is instantiated via the from_path() method, the basic
+    axes are populated with unique values from the filenames of the atmosphere
+    fits files. Associated axes are populated from the fits file contents. The
+    axes are then exported as numpy arrays. The model atmosphere can also be
+    extended with additional associated axes (for example, `ebvs` and `rvs` for
+    extincted intensities).
+
+    Attributes that are automatically populated:
+
+    * `basic_axes` (tuple): tuple of numpy arrays for basic axes
+    * `associated_axes` (tuple): tuple of numpy arrays for associated axes
+
+    In the case of from_path() instantiation, the following attributes are
+    also populated:
+
+    * `models` (list): list of atmosphere fits files
+    * `nmodels` (int): number of atmosphere fits files
+    * `indices` (array): array of indices for all defined nodes in the model
+        atmosphere
+    * `[axis_name]` (array): numpy array for each axis, where the name is
+        automatically inferred from the basic and associated axis names
+
+    Arguments
+    ----------
+    * `basic_axes` (tuple of ndarrays): values of the basic axes
+    * `associated_axes` (tuple of ndarrays): values of the associated axes
+    * `from_path` (bool): if True, the class is instantiated from a path
+
+    Raises
+    -------
+    * `FileNotFoundError`: if the path does not exist
+    """
+
+    name = None
+    prefix = None
+
+    # default axes:
+    basic_axis_names = ['teffs', 'loggs', 'abuns']
+    associated_axis_names = ['mus']
+
+    def __init__(self, basic_axes=None, associated_axes=None, from_path=False):
+        if from_path:
+            return
+
+        self.path = None
+
+        if basic_axes is None:
+            raise ValueError('basic_axes must be defined.')
+        if associated_axes is None:
+            raise ValueError('associated_axes must be defined.')
+
+        self.basic_axes = basic_axes
+        self.associated_axes = associated_axes
+
+    def register(self):
+        """
+        Registers the model atmosphere with the global model table.
+        """
+
+        global _atmtable
+        _atmtable.append(self.__class__)
+
+    @classmethod
+    def from_path(cls, path):
+        """
+        Instantiates the class and all attributes from a given path.
+
+        Arguments
+        ----------
+        * `path` (string): relative or absolute path to data files
+
+        Returns
+        --------
+        * a new model atmosphere object
+        """
+
+        self = cls(from_path=True)
+        self.path = path
+
+        try:
+            self.models = glob.glob(os.path.join(path, '*fits'))
+            self.nmodels = len(self.models)
+        except FileNotFoundError:
+            raise FileNotFoundError(f'path {path} does not exist.')
+
+        # initialize arrays for basic axes:
+        for name in self.basic_axis_names:
+            setattr(self, name, np.empty(self.nmodels))
+
+        # parse the filenames and populate the arrays:
+        for i, model in enumerate(self.models):
+            relative_filename = os.path.basename(model)
+            basic_node_values = self.parse_rules(relative_filename)
+            for j, name in enumerate(self.basic_axis_names):
+                getattr(self, name)[i] = basic_node_values[j]
+
+        # export basic and associated axes:
+        self.basic_axes = tuple([np.unique(getattr(self, name)) for name in self.basic_axis_names])
+        self.associated_axes = tuple([np.unique(getattr(self, name)) for name in self.associated_axis_names])
+
+        # store all node indices:
+        nodes = np.vstack([getattr(self, name) for name in self.basic_axis_names]).T
+        self.indices = np.empty_like(nodes, dtype=int)
+        for i, basic_axis in enumerate(self.basic_axes):
+            self.indices[:, i] = np.searchsorted(basic_axis, nodes[:, i])
+
+        return self
+
+    def parse_rules(self, relative_filename):
+        """
+        Provides rules for parsing atmosphere fits files containing data.
+        Only derived classes should implement this method.
+        """
+
+        return NotImplementedError
+
+    def add_associated_axis(self, name, axis):
+        """
+        Adds an associated axis to the model atmosphere.
+
+        Arguments
+        ----------
+        * `name` (string): name of the axis
+        * `axis` (array): values of the axis
+
+        Returns
+        --------
+        * a copy of the added axis.
+        """
+
+        self.associated_axis_names += [name,]
+        self.associated_axes += (axis,)
+        setattr(self, name, axis)
+
+    def remove_associated_axis(self, name):
+        """
+        Removes an associated axis from the model atmosphere.
+
+        Arguments
+        ----------
+        * `name` (string): name of the axis
+        """
+
+        if name in self.associated_axis_names:
+            idx = self.associated_axis_names.index(name)
+            self.associated_axis_names.pop(idx)
+            self.associated_axes = self.associated_axes[:idx] + self.associated_axes[idx+1:]
+            delattr(self, name)
+
+    def add_axis_node(self, axis_name, axis_node):
+        """
+        Adds a node to the specified axis. This method is used
+        when we want to add another value into the convex hull
+        spun by the current axes. If you do not know why you would
+        use this method, you likely should not use it.
+
+        Arguments
+        ----------
+        * `axis_name` (string): name of the axis
+        * `axis_node` (float): value of the node to be added
+        """
+
+        if axis_name in self.basic_axis_names:
+            new_axes = list(self.basic_axes)
+            axis_index = self.basic_axis_names.index(axis_name)
+            axis = self.basic_axes[axis_index]
+            if axis_node not in axis:
+                axis = np.append(axis, axis_node)
+                axis.sort()
+
+                new_axes[axis_index] = axis
+                self.basic_axes = tuple(new_axes)
+                nodes = np.vstack([getattr(self, name) for name in self.basic_axis_names]).T
+                self.indices = np.empty_like(nodes, dtype=int)
+                for i, basic_axis in enumerate(self.basic_axes):
+                    self.indices[:, i] = np.searchsorted(basic_axis, nodes[:, i])
+        else:
+            raise ValueError(f"Axis name '{axis_name}' not recognized.")
+
+    def limb_treatment(self, intensities):
+        """
+        Define how intensities at the exact limb (mu=0) should be treated. By
+        default, the intensities are linearly extrapolated to mu=0.
+
+        Arguments
+        ----------
+        * `intensities` (array): intensities across all mus
+
+        Returns
+        --------
+        * an array of intensities with the limb treatment applied.
+        """
+
+        intensities[0] = intensities[1] + (intensities[2]-intensities[1])/(self.mus[2]-self.mus[1])*(self.mus[0]-self.mus[1])
+        return intensities
+
+
+class BlackbodyModelAtmosphere(ModelAtmosphere):
+    """
+    Blackbody model atmosphere.
+
+    The blackbody model atmosphere is a simple model atmosphere that assumes
+    the object is a blackbody. The grid is defined by effective temperature
+    (teff).
+    """
+
+    name = 'blackbody'
+    prefix = 'bb'
+
+    basic_axis_names = ['teffs']
+    associated_axis_names = []
+    teffs = 10**np.linspace(2.5, 5.7, 97)  # this corresponds to the 316K-501187K range.
+
+    def __init__(self, *args, **kwargs):
+        super().__init__('blackbody', *args, **kwargs)
+
+    def limb_treatment(self, intensities):
+        return intensities
+
+
+class CK2004ModelAtmosphere(ModelAtmosphere):
+    """
+    Castelli & Kurucz (2004) model atmosphere.
+
+    The CK2004 model atmosphere is a grid of model atmospheres computed by
+    Castelli & Kurucz (2004). The grid is defined by effective temperature
+    (teff), surface gravity (logg), and chemical abundance (abun). The
+    intensities are computed for 37 angles (mus) on the 900-40000A wavelength
+    range.
+    """
+
+    name = 'ck2004'
+    prefix = 'ck'
+
+    mus = np.array([
+        0., 0.001, 0.002, 0.003, 0.005, 0.01, 0.015, 0.02, 0.025, 0.03, 0.035, 0.04,
+        0.045, 0.05, 0.06, 0.07, 0.08, 0.09, 0.1, 0.15, 0.2, 0.25, 0.3, 0.35, 0.4,
+        0.45, 0.5, 0.55, 0.6, 0.65, 0.7, 0.75, 0.8, 0.85, 0.9, 0.95, 1.
+    ])
+    wls = np.arange(900., 39999.501, 0.5)/1e10  # AA -> m
+    units = 1e7  # erg/s/cm^2/A -> W/m^3
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(atm=self.name, *args, **kwargs)
+
+    def parse_rules(self, relative_filename):
+        return [
+            float(relative_filename[1:6]),  # teff
+            float(relative_filename[7:9])/10,  # logg
+            float(relative_filename[10:12])/10 * (-1 if relative_filename[9] == 'M' else 1)  # abun
+        ]
+
+
+class PhoenixModelAtmosphere(ModelAtmosphere):
+    """
+    Phoenix (Husser et al. 2012) model atmosphere.
+
+    The Phoenix model atmosphere is a grid of model atmospheres computed by
+    Husser et al. (2012). The grid is defined by effective temperature (teff),
+    surface gravity (logg), and chemical abundance (abun). The intensities are
+    computed for 37 angles between 500 and 26000 Angstroms.
+    """
+
+    name = 'phoenix'
+    prefix = 'ph'
+
+    mus = np.array([
+        0., 0.001, 0.002, 0.003, 0.005, 0.01, 0.015, 0.02, 0.025, 0.03, 0.035, 0.04,
+        0.045, 0.05, 0.06, 0.07, 0.08, 0.09, 0.1, 0.15, 0.2, 0.25, 0.3, 0.35, 0.4,
+        0.45, 0.5, 0.55, 0.6, 0.65, 0.7, 0.75, 0.8, 0.85, 0.9, 0.95, 1.
+    ])
+    wls = np.arange(500., 26000.)/1e10  # AA -> m
+    units = 1  # W/m^3
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(self.name, *args, **kwargs)
+
+    def parse_rules(self, relative_filename):
+        return [
+            float(relative_filename[1:6]),  # teff
+            float(relative_filename[7:11]),  # logg
+            float(relative_filename[12:16])  # abun
+        ]
+
+
+class TMAPModelAtmosphere(ModelAtmosphere):
+    """
+    TMAP model atmosphere.
+    """
+
+    name = 'tmap'
+    prefix = 'tm'
+
+    basic_axis_names = ['teffs', 'loggs']
+    associated_axis_names = ['mus']
+
+    mus = np.array([
+        0., 0.00136799, 0.00719419, 0.01761889, 0.03254691, 0.05183939, 0.07531619,
+        0.10275816, 0.13390887, 0.16847785, 0.20614219, 0.24655013, 0.28932435,
+        0.33406564, 0.38035639, 0.42776398, 0.47584619, 0.52415388, 0.57223605,
+        0.6196437, 0.66593427, 0.71067559, 0.75344991, 0.79385786, 0.83152216,
+        0.86609102, 0.89724188, 0.92468378, 0.9481606,  0.96745302, 0.98238112,
+        0.99280576, 0.99863193, 1.
+    ])
+    units = 1  # W/m^3
+
+    def __init__(self, *args, **kwargs):
+        super().__init__('tmap', *args, **kwargs)
+        self.wls = np.load(kwargs['path'] + '/wavelengths.npy')  # in meters
+
+    def parse_rules(self, relative_filename):
+        pars = re.split('[TGA.]+', relative_filename)
+        return [
+            float(pars[1]),  # teff
+            float(pars[2])/100  # logg
+        ]
+
+
+# global model atmosphere table:
+_atmtable = ModelAtmosphere.__subclasses__()
